Updating the ui and extracting
